Calendar
The biggest update to the calendar since, well, the calendar. Using the open-source fullcalendar.io technologies to create a working Google calendar javascript piece of art. I don't know javascript at the moment so it has taken me hours to modify and get the API working for this piece.

Known bugs:
-overlaps the navigation bar
-appears cramped on mobile
-height needs to be adjusted

Not bugs but major problem:
-Blue valley has a very messy and inaccurate district calendar that I cannot fix, as it always is chaning. I am at the mercy of BVNW